store: Move AdaptiveBatchSize to its own module
